move ogp validation codes to ogp route
