+++ packages/app/src/server/routes/ogp.ts
@@ -1,6 +1,7 @@
 import {
-  Request, Response,
+  Request, Response, NextFunction,
 } from 'express';
+import { param, validationResult, ValidationError } from 'express-validator';
 
 import path from 'path';
 import * as fs from 'fs';
@@ -50,60 +51,89 @@ module.exports = function(crowi) {
 
   };
 
-  return {
-    async renderOgp(req: Request, res: Response) {
+  const renderOgp = async(req: Request, res: Response) => {
 
-      const { configManager } = crowi;
-      const ogpUri = configManager.getConfig('crowi', 'app:ogpUri');
+    const { configManager } = crowi;
+    const ogpUri = configManager.getConfig('crowi', 'app:ogpUri');
 
-      const pageId = req.params.pageId;
-      let user;
-      let pageTitle: string;
-      let bufferedUserImage: Buffer;
+    const pageId = req.params.pageId;
+    let user;
+    let pageTitle: string;
+    let bufferedUserImage: Buffer;
 
-      try {
-        const Page = crowi.model('Page');
-        const page = await Page.findByIdAndViewer(pageId);
+    try {
+      const Page = crowi.model('Page');
+      const page = await Page.findByIdAndViewer(pageId);
 
-        if (page == null || page.status !== Page.STATUS_PUBLISHED || (page.grant !== Page.GRANT_PUBLIC && page.grant !== Page.GRANT_RESTRICTED)) {
-          return res.status(400).send('the page does not e xist');
-        }
-
-        const User = crowi.model('User');
-        user = await User.findById(page.creator._id.toString());
-
-        bufferedUserImage = user.imageUrlCached === DEFAULT_USER_IMAGE_URL ? bufferedDefaultUserImageCache : (await getBufferedUserImage(user.imageUrlCached));
-        // todo: consider page title
-        pageTitle = (new DevidedPagePath(page.path)).latter;
-      }
-      catch (err) {
-        logger.error(err);
-        return res.status(500).send(`error: ${err}`);
+      if (page == null || page.status !== Page.STATUS_PUBLISHED || (page.grant !== Page.GRANT_PUBLIC && page.grant !== Page.GRANT_RESTRICTED)) {
+        return res.status(400).send('the page does not exist');
       }
 
-      let result;
-      try {
-        result = await axios.post(
-          ogpUri, {
-            data: {
-              title: pageTitle,
-              userName: user.username,
-              userImage: bufferedUserImage,
-            },
-          }, {
-            responseType: 'stream',
+      const User = crowi.model('User');
+      user = await User.findById(page.creator._id.toString());
+
+      bufferedUserImage = user.imageUrlCached === DEFAULT_USER_IMAGE_URL ? bufferedDefaultUserImageCache : (await getBufferedUserImage(user.imageUrlCached));
+      // todo: consider page title
+      pageTitle = (new DevidedPagePath(page.path)).latter;
+    }
+    catch (err) {
+      logger.error(err);
+      return res.status(500).send(`error: ${err}`);
+    }
+
+    let result;
+    try {
+      result = await axios.post(
+        ogpUri, {
+          data: {
+            title: pageTitle,
+            userName: user.username,
+            userImage: bufferedUserImage,
           },
-        );
-      }
-      catch (err) {
-        logger.error(err);
-        return res.status(500).send(`error: ${err}`);
-      }
+        }, {
+          responseType: 'stream',
+        },
+      );
+    }
+    catch (err) {
+      logger.error(err);
+      return res.status(500).send(`error: ${err}`);
+    }
 
-      res.writeHead(200, {
-        'Content-Type': 'image/jpeg',
-      });
-      result.data.pipe(res);
-    },
+    res.writeHead(200, {
+      'Content-Type': 'image/jpeg',
+    });
+    result.data.pipe(res);
+
   };
+
+  const pageIdRequired = param('pageId').not().isEmpty().withMessage('page id is not included in the parameter');
+
+  const ogpValidator = async(req:Request, res:Response, next:NextFunction) => {
+    const { aclService, fileUploadService, configManager } = crowi;
+
+    const ogpUri = configManager.getConfig('crowi', 'app:ogpUri');
+
+    if (ogpUri == null) return res.status(400).send('OGP URI for GROWI has not been setup');
+    if (!fileUploadService.getIsUploadable()) return res.status(501).send('This GROWI can not upload file');
+    if (!aclService.isGuestAllowedToRead()) return res.status(501).send('This GROWI is not public');
+
+    const errors = validationResult(req);
+
+    if (errors.isEmpty()) {
+      return next();
+    }
+
+    // errors.array length is one bacause pageIdRequired is used
+    const pageIdRequiredError: ValidationError = errors.array()[0];
+
+    return res.status(400).send(pageIdRequiredError.msg);
+  };
+
+  return {
+    renderOgp,
+    pageIdRequired,
+    ogpValidator,
+  };
+
 };
